Remove MIT License tag from Banana Cake Pop (#55962)
* Remove MIT License tag from Banana Cake Pop

Banana Cake Pop is free to use but not open source and not under the MIT license. 
The license file that was referred to was only covering the Hot Chocolate project. 
Banana Cake Pop is not part of that repository.

I have remove MIT License tag and rhe reference to the license file of Hot Chocolate.

* Update manifests/c/ChilliCream/BananaCakePop/1.0.0-preview.31/ChilliCream.BananaCakePop.locale.en-US.yaml

Co-authored-by: Esco <OfficialEsco@users.noreply.github.com>

Co-authored-by: Esco <OfficialEsco@users.noreply.github.com>